requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==2.1.7
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
google-auth==1.16.1
idna==2.9
kubernetes==11.0.0
oauthlib==3.1.0
Pillow==7.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
websocket-client==0.57.0
